Consolidated all delayed code into a single file.
